Move import tests to a distinct suite
